Fix observation space: enhance switch_states with locked door info
- Enhanced switch_states from 5 to 30 features (6 per door * 5 doors)
- Per-door format: [switch_x, switch_y, door_x, door_y, collected, open]
- Added detailed locked door tracking with normalization
- Fixed observation space consistency issues
- Made reachability_features always present
- Fixed subtask_features not appearing in processed observations
- Fixed game_state dtype mismatch (float64 vs float32)
- All observation validation tests passing

+++ nclone/gym_environment/observation_processor.py
@@ -446,12 +446,13 @@ class ObservationProcessor:
         # For backward compatibility, if we have entity states, include them
         # but the new design focuses on the enhanced ninja state
         if len(obs["game_state"]) > 26:
-            entity_states = obs["game_state"][26:]
+            entity_states = obs["game_state"][26:].astype(np.float32)  # Ensure float32
             processed_state = np.concatenate([ninja_state, entity_states])
         else:
             processed_state = ninja_state
 
-        return processed_state
+        # Ensure final state is float32
+        return processed_state.astype(np.float32)
 
     def process_rendered_global_view(self, screen: np.ndarray) -> np.ndarray:
         """Process the rendered screen into a downsampled global view.
